Xcode8: fix relative line numbers
