+++ XVim/DVTTextSidebarView+XVim.m
@@ -25,15 +25,52 @@ static CGFloat kTextSideBarLineNumberRightPadding = 5.0;
 
 @implementation DVTTextSidebarView(XVim)
 + (void)xvim_initialize{
-    [self xvim_swizzleInstanceMethod:@selector(_drawLineNumbersInSidebarRect:foldedIndexes:count:linesToInvert:linesToReplace:getParaRectBlock:) with:@selector(xvim__drawLineNumbersInSidebarRect:foldedIndexes:count:linesToInvert:linesToReplace:getParaRectBlock:)];
+#ifdef XVIM_XCODE8
+    [self xvim_swizzleInstanceMethod:@selector(_drawLineNumbersInSidebarRect:foldedIndexes:count:linesToInvert:linesToHighlight:linesToReplace:textView:getParaRectBlock:)
+                                with:@selector(xvim__drawLineNumbersInSidebarRect:foldedIndexes:count:linesToInvert:linesToHighlight:linesToReplace:textView:getParaRectBlock:)];
+#else
 
+    [self xvim_swizzleInstanceMethod:@selector(_drawLineNumbersInSidebarRect:foldedIndexes:count:linesToInvert:linesToReplace:getParaRectBlock:)
+                                with:@selector(xvim__drawLineNumbersInSidebarRect:foldedIndexes:count:linesToInvert:linesToReplace:getParaRectBlock:)];
+#endif
 }
+
+// Xcode 8
+- (void)xvim__drawLineNumbersInSidebarRect:(CGRect)arg1
+                             foldedIndexes:(NSUInteger *)arg2
+                                     count:(NSUInteger)arg3
+                             linesToInvert:(id)arg4
+                          linesToHighlight:(id)arg5
+                            linesToReplace:(id)arg6
+                                  textView:(id)arg7
+                          getParaRectBlock:(GetParaBlock)arg8
+{
+        if (XVim.instance.options.relativenumber) {
+                for (NSUInteger i = 0 ; i < arg3 ; ++i) {
+                        unsigned long long lineNumber = arg2[i];
+                        [self xvim_drawRelativeNumberForLineNumber:lineNumber];
+                }
+        }
+        else {
+                [self xvim__drawLineNumbersInSidebarRect:(CGRect)arg1
+                                           foldedIndexes:(NSUInteger *)arg2
+                                                   count:(NSUInteger)arg3
+                                           linesToInvert:(id)arg4
+                                        linesToHighlight:(id)arg5
+                                          linesToReplace:(id)arg6
+                                                textView:(id)arg7
+                                        getParaRectBlock:(GetParaBlock)arg8];
+        }
+}
+
+// Pre-Xcode 8
 - (void)xvim__drawLineNumbersInSidebarRect:(struct CGRect)arg1
                         foldedIndexes:(unsigned long long *)arg2
                                 count:(unsigned long long)arg3
                         linesToInvert:(id)arg4
                        linesToReplace:(id)arg5
-                     getParaRectBlock:(id)arg6{
+                     getParaRectBlock:(id)arg6
+{
     
     if (XVim.instance.options.relativenumber) {
         for (NSUInteger i = 0 ; i < arg3 ; ++i) {
@@ -50,20 +87,21 @@ static CGFloat kTextSideBarLineNumberRightPadding = 5.0;
     }
 }
 
-- (void)xvim_drawRelativeNumberForLineNumber:(unsigned long long)lineNumber{
-       
+- (void)xvim_drawRelativeNumberForLineNumber:(NSUInteger)lineNumber
+{
     struct CGRect paraRect;
     struct CGRect firstLineRect;
-    
+        
     [self getParagraphRect:&paraRect firstLineRect:&firstLineRect forLineNumber:lineNumber];
     
     DVTSourceTextView *sourceTextView = self.scrollView.documentView;
-    long long currentLineNumber = [sourceTextView _currentLineNumber];
-    long long relativeLineNumber = llabs(((long long)lineNumber - currentLineNumber));
-    if (XVim.instance.options.number && relativeLineNumber == 0) relativeLineNumber = (long long)lineNumber;
-    NSString *relativeLineNumberString = [@(relativeLineNumber) stringValue];
+    NSUInteger relativeLineNumber = (NSUInteger)llabs(((long long)lineNumber - sourceTextView._currentLineNumber));
+    BOOL drawLineIsCurrentLine = (XVim.instance.options.number && relativeLineNumber == 0);
+    NSUInteger drawLineNumber =  drawLineIsCurrentLine ? lineNumber : relativeLineNumber;
+    NSString *relativeLineNumberString = [@(drawLineNumber) stringValue];
+    NSColor *drawLineNumberColor = drawLineIsCurrentLine ? [NSColor colorWithWhite:0.8 alpha:1.0] : [self lineNumberTextColor];
     
-    NSDictionary *attributes = @{NSForegroundColorAttributeName: [self lineNumberTextColor],
+    NSDictionary *attributes = @{NSForegroundColorAttributeName: drawLineNumberColor,
                                  NSFontAttributeName: [self lineNumberFont],
                                  NSParagraphStyleAttributeName: [NSParagraphStyle dvt_paragraphStyleWithAlignment:NSRightTextAlignment]};
     NSAttributedString *attributedString = [[NSAttributedString alloc] initWithString:relativeLineNumberString attributes:attributes];
